+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 ==================
 
   * Added Unit tests and test coverage information
-  * Added Travis CI support
+  * Save button is disabled until command and name are not empty
   * Removed "pointer" cursor style
   * Fixed bug where stopping a suite will turn all tasks icons to "stopped"
   * Fixed a small bug where empty paths where stored in the tasks.json file
+  * Added Travis CI support
   * Minor code refactoring